github-mcp-app/pyproject.toml
[project]
name = "github-mcp-server"
version = "0.1.0"
description = "GitHub MCP Server for DataScope - searches novatech-transformations repo"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "databricks-sdk>=0.20.0",
    "PyGithub>=2.1.0",
]

[project.scripts]
github-mcp-server = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
